==> R/1-data-extraction.Rmd <==
---
title: "New Manager Bounce"
subtitle: "Data Extraction"
author: "Josh Muncke"
date: "20/03/2019"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Data Extraction

The first step in the process is to extract the datasets needed for this analysis from remote sources.

There are three primary datasets required for this analysis:

1. The match data (results per EPL match) - from [football-data.co.uk](http://football-data.co.uk).
2. The manager data including start and end dates for each manager at each club from Wikipedia.
3. A mapping of club names between these two data sources - manually created.


We will be using `rvest` to scrape the two remote data sets. Let's get started!

## Setup 

Load the required R packages.

```{r load our packages, message=FALSE, warning=FALSE}
library(tidyverse)
library(rvest)
library(magrittr)
library(feather)
```

## Match data

### Extraction

We're going to scrape the game data from the excellent [football-data.co.uk](http://football-data.co.uk) site. The game-by-game data files are stored like this: `http://football-data.co.uk/mmz4281/1819/E0.csv` where the `1819` signifies the 2018-2019 season.

To get at this we'll create a vector of game files and then `purrr::map` the `read_csv` function over it. Sweet.

```{r game data files, message=FALSE}
firstyear <- 1995
lastyear <- 2019

from_date <- seq(lastyear - 1, firstyear, -1) %>% as.character() %>% substr(3,4)
to_date <- seq(lastyear, firstyear + 1, -1) %>% as.character() %>% substr(3,4)
date_seq <- paste0(from_date, to_date)

game_data_filenames <- paste0("http://football-data.co.uk/mmz4281/", date_seq, "/E0.csv")
```

Now download the files and `reduce` them into a single data frame.

```{r download game data, message=FALSE, warning=FALSE}

games_raw <- game_data_filenames %>% 
  map(function(x) read_csv(x) %>% select(Date:FTAG) %>% as.data.frame()) %>% 
  reduce(rbind) %>%
  as_tibble() %>%
  filter(complete.cases(.)) %>% 
  mutate(Date = as.character(Date)) %>%
  as_tibble()
```

There's a bunch of `problems()` with this file because of column inconsistencies but the main thing we need for our analysis - dates, teams and results are read in okay. Let's shut our eyes on the other stuff and pretend we didn't see it.

### Saving

Now we save this raw data locally.

```{r save game data}
games_raw %>%
  write_feather("../data/raw/game_data_raw.feather")
```

## Manager Data

### Extractions

Now we will pull the list of EPL managers, clubs and time-periods from the Wikipedia article on [Premier League Managers](https://en.wikipedia.org/wiki/List_of_Premier_League_managers).

We're going to use `extract2` from the `magrittr` library which is a nice pipe-able replacement for `[[`. The actually scraping is done by `rvest`.

```{r scrape manager data}
url <- "https://en.wikipedia.org/wiki/List_of_Premier_League_managers"

managers_raw <- url %>% 
  read_html() %>% 
  html_nodes("table") %>% 
  extract2(2) %>% 
  html_table(fill = T) %>% 
  as_tibble()
```

### Saving

Write the manager data locally.

```{r save manager data}
managers_raw %>%
  write_feather("../data/raw/manager_data_raw.feather")
```

## Club name mapping

### Definition

Between the two datasets some clubs are named differently (e.g. "Man United" vs "Manchester United"). I considered doing a little string similarity comparison to fix this but ultimately I took the lazy/boring route and just created a separate lookup table.

```{r club name matching}
club_name_mapping <- tribble(
  ~short_name, ~long_name,
  "Man United", "Manchester United",
  "Huddersfield", "Huddersfield Town",
  "Newcastle", "Newcastle United",
  "Wolves", "Wolverhampton Wanderers",
  "Cardiff", "Cardiff City",
  "Leicester", "Leicester City",
  "Tottenham", "Tottenham Hotspur",
  "West Ham", "West Ham United",
  "Brighton", "Brighton & Hove Albion",
  "Man City", "Manchester City",
  "Bolton", "Bolton Wanderers",
  "Blackburn", "Blackburn Rovers",
  "Wigan", "Wigan Athletic",
  "QPR", "Queens Park Rangers",
  "Hull", "Hull City",
  "Swansea", "Swansea City",
  "Stoke", "Stoke City",
  "West Brom", "West Bromwich Albion",
  "Derby", "Derby County",
  "Charlton", "Charlton Athletic",
  "Leeds", "Leeds United",
  "Middlesboro", "Middlesbrough",
  "Ipswich", "Ipswich Town",
  "Coventry", "Coventry City",
  "Bradford", "Bradford City",
  "Sheffield Weds", "Sheffield Wednesday",
  "Nott'm Forest", "Nottingham Forest",
  "Norwich", "Norwich City",
  "Birmingham", "Birmingham City")
```

### Saving

Write the mapping data locally.

```{r save mapping data}
club_name_mapping %>%
  write_feather("../data/raw/mapping_data_raw.feather")
```

## Summary

We have now extracted the raw data we require for this analysis. In the next file we will cleanse, prepare and combine into a single file for modelling!

```{r session info}
sessionInfo()
```
