==> R/2-data-preparation.Rmd <==
---
title: "New Manager Bounce"
subtitle: "Data Preparation"
author: "Josh Muncke"
date: "20/03/2019"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Data Preparation

Step 2 is to cleanse, prepare and join the manager and game data.

## Setup 

Load the required R packages.

```{r load our packages, message=FALSE, warning=FALSE}
library(tidyverse)
library(magrittr)
library(feather)
library(lubridate)
library(testthat)
library(redbull)
```

Load the raw extracted data into memory.

```{r load raw data}
games_raw <- read_feather("../data/raw/game_data_raw.feather")
managers_raw <- read_feather("../data/raw/manager_data_raw.feather")
club_names <- read_feather("../data/raw/mapping_data_raw.feather")
```

## Game data

### Replace club names

The names of teams between the manager and game data is slightly different for some clubs (e.g. "Man Utd" vs "Manchester United"). In order to remedy this we will use the mapping table we created in the first step to replace the team names for both the home and away team columns of each game.

```{r map club names}
games_prepared <- games_raw

games_prepared %<>% 
  left_join(club_names, by = c("HomeTeam" = "short_name")) %>%
  mutate(HomeTeam = ifelse(is.na(long_name), HomeTeam, long_name)) %>%
  select(-long_name) %>% 
  left_join(club_names, by = c("AwayTeam" = "short_name")) %>%
  mutate(AwayTeam = ifelse(is.na(long_name), AwayTeam, long_name)) %>%
  select(-long_name)
```

### Expand game data

The raw game data has one row per match with the results for both the home and away teams. Since we want to do our analysis at a club-by-club level we effectively need to double this data so that we have a single row per team per game indicating who the opponent was, the scores for each and whether that team was playing at home or away.

We first add an identifier column (just the row number) for each game so we can keep track of missing/duplicates later on.

```{r add game id}
games_prepared %<>% 
  select(game_date = Date, everything()) %>% 
  mutate(game_id = row_number())
```

We now update the game date column to format it to an actual date (there's a couple of different date formats to work with).

```{r update game date}
games_prepared %<>% 
  mutate(game_date = case_when(
  str_length(game_date) == 10 ~ as_date(game_date, format = "%d/%m/%Y", tz = "UTC"),
  str_length(game_date) == 8 ~ as_date(game_date, format = "%d/%m/%y", tz = "UTC"),
  TRUE ~ as_date('1900-01-01', tz = "UTC")))
```

And finally we create two temporary versions of this data - one for the home result and one for the away result - before binding them together.

We also make sure to create a "points" column indicating the number of points the team received for the match (0 for a loss, 1 for a draw, 3 for a win).

```{r create results data}
home_games <- games_prepared %>% 
  select(game_id, game_date, game_team = HomeTeam, goals = FTHG, opponent = AwayTeam, opponent_goals = FTAG) %>%
  mutate(game_location = "Home", 
         points = case_when(goals > opponent_goals ~ 3, goals < opponent_goals ~ 0, TRUE ~ 1))

away_games <- games_prepared %>% 
  select(game_id, game_date, game_team = AwayTeam, goals = FTAG, opponent = HomeTeam, opponent_goals = FTHG) %>%
  mutate(game_location = "Away", 
         points = case_when(goals > opponent_goals ~ 3, goals < opponent_goals ~ 0, TRUE ~ 1))

game_results_full <- bind_rows(home_games, away_games)

rm(list = c('home_games','away_games'))
```

### Test and visualize

A good test would be to make sure we have the right number of expected games between the `full` and `prepared` data sets and also check we don't have any games that have greater or less than two records.

We're going to use the `testthat` package to check this.

```{r compare game ids}
test_that('Same number of distinct game ids',
          expect_equal(
            games_prepared %>% distinct(game_id) %>% tally(),
            game_results_full %>% distinct(game_id) %>% tally()
          ))
```

```{r check only two records per game}
test_that('Same number of distinct game ids',
          expect_equal(
            game_results_full %>% add_count(game_id) %>% filter(n != 2) %>% tally() %>% pull(nn),
            0
          ))
```

Finally let's do a quick analysis to look at the average number of games per month over time - just a simple visual gut check. I'm going to use `ggplot2` and shamelessly self-promote the `[redbull](http://github.com/deathbydata/redbull)` package I developed for themeing.

```{r visualize games over time, warning=FALSE}
theme_set(theme_rb())

game_results_full %>%
  mutate(game_month = round_date(game_date, "month")) %>%
  group_by(game_month) %>%
  summarize(n_games = n_distinct(game_id)) %>%
  ggplot(aes(x = game_month, y = n_games)) +
  geom_line() +
  labs(title = "Number of EPL games per month since 1995",
       caption = "Source: football-data.co.uk",
       x = "Month",
       y = "Number of games")
```

All looks good so we write this prepared game data back locally.

```{r save game data}
game_results_full %>%
  write_feather("../data/prep/game_data_prep.feather")
```

```{r remove un-needed dfs}
rm(list = c('club_names', 'games_prepared', 'games_raw'))
```

## Manager data

### Basic cleaning

Like most data that is scraped from a web-page the manager data is a little ugly and requires some love.

For our purposes we only require the name of the manager, the club they managed and the time window so we `select` these columns and drop the rest.

```{r prepare manager data}
managers_prepared <- managers_raw

managers_prepared %<>% select(manager_name = Name, 
                          managed_club = Club, 
                          managed_from = From, 
                          managed_until = Until)
```

Some of the date data needs to be fixed. In particular for incumbent managers the `managed_until` column has the value "Present*" which we will replace with today's date.

We also coerce the string dates into a formatted date field.

```{r fix manager dates}
managers_prepared %<>% 
  mutate(managed_until = if_else(managed_until == 'Present*', format.Date(Sys.Date() + 1, format = "%e %B %Y"), managed_until))

managers_prepared %<>% mutate(managed_from = lubridate::as_date(managed_from, format = "%e %B %Y", tz = "UTC"), 
                          managed_until = lubridate::as_date(managed_until, format = "%e %B %Y", tz = "UTC"))
```

### Fixing manager date ranges

In an ideal world this manager data would cover all timeframes for all clubs and would map a single manager to each club at any given point of time. As you might expect - this is not the case.

There are instances of multiple managers managing the same club simultaneously (e.g. Ray Lewington and Ron Noades at Crystal Palace in 1998); of clubs which have a period of time with **no** manager (e.g. there was a [two-week gap](https://www.worldfootball.net/teams/leicester-city/9/) between Nigel Pearson leaving Leicester City on 29th June 2015 and the appointment of Claudio Ranieri on 13th July 2015; and finally of plain erroneous data in the Wikipedia table (e.g. Tomas Brolin didn't manage Crystal Palace in 1998 - although [he did play there](https://www.worldfootball.net/teams/crystal-palace/9/)).

Fixing these issues requires some deep digging and also some judgement calls since our analysis will require that each match for each club can be attributed to a single manager. The errors we can remove after some research but the gaps and overlaps need thinking - who do we attribute games to if there were two managers? What about if there was no manager?

Whilst that's logically okay it doesn't really work for this analysis since it will causes duplicate results.

To help us find and iterate on the issues, we build a small function that will join the `game_results` and `managers` data and return any matches with issues (either missing or duplicated).

```{r find problem matches function}
get_problem_matches <- function() {
  match_manager_join <- game_results_full %>% 
    left_join(managers_prepared, by = c('game_team' = 'managed_club')) %>%
    filter(game_date >= managed_from, game_date <= managed_until)
  
  missing_games <- game_results_full %>% anti_join(match_manager_join, by = "game_id") %>% mutate(problem = "missing")
  duplicated_games <- match_manager_join %>% add_count(game_id) %>% filter(n > 2) %>% mutate(problem = "duplicated")
  partial_games <- match_manager_join %>% group_by(game_id) %>% summarize(n_teams = n_distinct(game_team), n_rows = n()) %>% filter(n_teams != 2 | n_rows < 2) %>% select(game_id) %>% inner_join(match_manager_join, by = "game_id") %>% mutate(problem = "partial")
  
  errors <- duplicated_games %>% bind_rows(partial_games) %>% bind_rows(missing_games) %>% arrange(game_team, game_id, desc(game_date))
  
  errors
}

get_problem_matches() %>% View()
```

```{r}
get_problem_matches() %>% distinct(game_id) %>% tally()
get_problem_matches() %>% 
  group_by(manager_name, game_team, managed_from, managed_until) %>% 
  summarize(n_games = n_distinct(game_id)) %>% 
  arrange(game_team, managed_from) %>%
  View()
```

After lots of iteration we can identify the following fixes to be applied to the manager data:

1. Remove Tomas Brolin as ever having managed Crystal Palace (7 games) - he was just a player
2. Remove Adam Sadler as Leicester Manager (1 game) - he was caretaker ([data](https://www.worldfootball.net/teams/leicester-city/9/)).
3. Fix Brendan Rodgers' start date at Leicester from 26th Feb - 27th Feb 2019. (1 game)
4. Change Gerard Houllier's start date at Liverpool to 12th Nov 1998 and end date to 30th June 2004 (12 games)
5. Remove John Deehan as Aston Villa manager (only caretaker) - (4 games)
6. Modify Gérard Houllier's end date at Aston Villa to 22nd April 2011 - he was taken ill on 20th April [did not return for any subsequent matches](https://en.wikipedia.org/wiki/G%C3%A9rard_Houllier#Aston_Villa) - (5 games)
7. Rémi Garde started at Aston Villa on 2nd Nov 2015, fix Kevin MacDonald's end date to be 1st Nov 2015 - (1 game)

```{r manager cleanup}
managers_prepared %<>%
  filter(! (manager_name == "Tomas Brolin" & managed_club == "Crystal Palace")) %>%
  filter(! (manager_name == "Adam Sadler" & managed_club == "Leicester City")) %>%
  mutate(managed_from = if_else((manager_name == "Brendan Rodgers" & managed_club == "Leicester City"), as.Date("2019-02-27"), managed_from)) %>%
  filter(! (manager_name == "Phil Thompson" & managed_club == "Liverpool")) %>%
  filter(! (manager_name == "Gérard Houllier" & managed_club == "Liverpool" & managed_from == as.Date('2002-03-17'))) %>%
  mutate(managed_from = if_else((manager_name == "Gérard Houllier" & managed_club == "Liverpool"), as.Date("1998-11-12"), managed_from)) %>%
  mutate(managed_until = if_else((manager_name == "Gérard Houllier" & managed_club == "Liverpool"), as.Date("2004-06-30"), managed_until)) %>%
  filter(! (manager_name == "John Deehan" & managed_club == "Aston Villa")) %>%
  mutate(managed_until = if_else((manager_name == "Gérard Houllier" & managed_club == "Aston Villa"), as.Date("2011-04-22"), managed_until)) %>%
  mutate(managed_until = if_else((manager_name == "Kevin MacDonald" & managed_club == "Aston Villa" & managed_until == as.Date("2015-11-02")), as.Date("2015-11-01"), managed_until)) #%>%


# %>%
#   filter(! (manager_name == "Ray Lewington" & managed_club == "Crystal Palace")) %>%
#   filter(! (manager_name == "Mike Stowell" & managed_club == "Leicester City")) %>%
#   filter(! (manager_name == "Kevin Bond" & managed_club == "Queens Park Rangers")) %>%
#   mutate(managed_until = if_else((manager_name == "Roy Evans" & managed_club == "Liverpool" & managed_until == as.Date("1998-11-12")), as.Date("1998-06-30"), managed_until))

```

Also - some of the management date ranges are a little tricky. In some cases the new manager takes over the day *after* the old manager (this is fine). In other cases it's the *same* day (this causes problems). In yet other cases it is many days later (could also causes problems).

This all requires manual investigation and fixing.

```{r fix manager date ranges}
managers_prepared %<>%
  arrange(managed_club, managed_from) %>%
  group_by(managed_club) %>%
  mutate(next_mgr_started = lead(managed_from)) %>% 
  mutate(overlap = as.numeric(next_mgr_started - managed_until)) %>%
  mutate(overlap = if_else(is.na(overlap), 1, overlap)) %>%
  mutate(managed_until = if_else(overlap == 0, managed_until - 1, managed_until)) %>%
  dplyr::select(-next_mgr_started, -overlap) %>%
  ungroup()
```

Finally save the prepared manager data.

```{r save prepared manager data}
managers_prepared %>%
  write_feather("../data/prep/manager_data_prep.feather")
```

## Join results to managers by club name

Use the club name to join results to managers and remove any matches which do not fall between the manager start/end dates.

```{r join results}
game_and_manager_data_prepared <- game_results_full %>%
  left_join(managers_prepared, by = c("game_team" = "managed_club")) %>%
  filter(game_date >= managed_from & game_date <= managed_until)
```

```{r filter valid games}
game_and_manager_data_prepared %<>%
  add_count(game_id) %>% 
  filter(n == 2) %>%
  select(-n)
```

Write to feather!

```{r write final game and manager data}
game_and_manager_data_prepared %>% write_feather("../data/prep/game_and_manager_data_prep.feather")
```
