==> R/3-exploratory.Rmd <==
---
title: "New Manager Bounce"
subtitle: "Exploratory Analysis"
author: "Josh Muncke"
date: "21/03/2019"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Exploratory Analysis

```{r load our packages, message=FALSE}
library(tidyverse)
library(lubridate)
library(magrittr)
library(redbull)
library(feather)
```

## Load data

```{r load data}
game_and_manager_data <- read_feather("../data/prep/game_and_manager_data_prep.feather")
```

## Plotting setup

```{r plot setup}
theme_set(theme_rb())
source_text <- "Source: football-data.co.uk; Wikipedia"
```

## Exploratory Analysis

### Simple questions

**Average goals per game over time**
```{r gpg over time, warning=F}
game_and_manager_data %>%
  mutate(game_month = round_date(game_date, "month")) %>%
  group_by(game_month) %>%
  summarise(goals = sum(goals), games = n_distinct(game_id)) %>%
  mutate(goals_per_game = goals/games) %>%
  ggplot(aes(x = game_month, y = (goals_per_game))) +
  geom_point() + 
  geom_smooth(method = "loess", colour = "red", se = F) + 
  labs(caption = source_text,
       title = "Increasing number of goals per game",
       x = "Game Month",
       y = "Average Goals / Game")
```

**Distribution of goals per team per game**
```{r}
game_and_manager_data %>%
  ggplot(aes(x = goals, fill = game_location)) +
  geom_histogram(binwidth = 1) +
  facet_wrap(~ game_location) +
  scale_fill_redbull("bloomberg") +
  labs(caption = source_text,
      title = "Home turf advantage!",
     subtitle = "Distribution of goals scored home and away",
     x = "Goals",
     y = "Number of games")
```
## Tenure

### New features
```{r tenure features}
game_and_manager_data %<>%
  mutate(mgr_tenure_at_game_in_weeks = as.numeric(game_date - managed_from)/7) %>%
  mutate(mgr_time_until_end_in_weeks = as.numeric(managed_until - game_date)/7) %>%
  mutate(goal_difference = goals - opponent_goals)
```

**Correlation between goals scored by team and manager tenure at time of match**

```{r tenure goals relationship}
game_and_manager_data %>%
  filter(mgr_time_until_end_in_weeks <= 2 * 52) %>%
  ggplot(aes(x = mgr_time_until_end_in_weeks, y = jitter(goals), color = game_location)) +
  geom_point(alpha = 0.005) +
  facet_wrap(~ game_location) +
  scale_color_redbull("bloomberg") +
  geom_smooth(method = "loess") +
  scale_x_reverse()
```

```{r}

goals_model <- glm(goals ~ game_month + game_location, 
                   family = poisson(link = "log"), 
                   data = full_results %>% 
                     filter(! game_team %in% c('Manchester United', 'Arsenal', 'Charlton Athletic')) %>% 
                     mutate(manager_tenure = manager_tenure/365) %>%
                     mutate(game_month = as.factor(month(game_date))))

summary(goals_model)

```

```{r}
full_results %>%
  #filter(!manager_name %in% c('Alex Ferguson', 'Arsène Wenger', 'Alan Curbishley')) %>%
  mutate(weeks_remaining = floor(manager_time_remaining)) %>%
  filter(weeks_remaining <= 104) %>%
  group_by(weeks_remaining, game_location) %>%
  summarize(average_points = mean(points), ngames = n_distinct(game_id)) %>%
  ggplot(aes(x = weeks_remaining, y = average_points, color = game_location, size = ngames)) +
  geom_point(alpha = 0.7) +
  scale_x_reverse() +
  ylim(0.5,2.5) +
  geom_smooth(method = "lm", formula = y ~ poly(x, 4), se = F) +
  geom_hline(yintercept = mean(full_results$points))

full_results %>%
  #filter(!manager_name %in% c('Alex Ferguson', 'Arsène Wenger', 'Alan Curbishley')) %>%
  mutate(weeks_tenure = floor(manager_tenure)) %>%
  filter(weeks_tenure <= 104) %>%
  group_by(weeks_tenure, game_location) %>%
  summarize(average_points = mean(points), ngames = n_distinct(game_id)) %>%
  ggplot(aes(x = weeks_tenure, y = average_points, color = game_location)) +
  geom_point(alpha = 0.7) +
  #ylim(0.5,2.5) +
  geom_smooth(method = "gam", formula = y ~ splines::ns(x, 4), se = F) +
  facet_wrap(~game_location,scales = "free") +
  geom_hline(yintercept = mean(full_results$points)) 
```

```{r}
full_results %>%
  filter(game_team == "Chelsea") %>%
  ggplot(aes(x = game_date, y = points, colour = paste(manager_name, managed_from, sep = ","))) +
  geom_point(alpha = 0.1) +
  geom_smooth(method = "lm", formula = y ~ poly(x, 3), se = F)

```
```



  #filter(manager_tenure <= 8*365.25) %>%
  ggplot(aes(x = manager_time_remaining/365.25, y = jitter(points))) +
  geom_point(alpha = 0.3) +
  geom_smooth() +
  facet_wrap(~game_location)

full_results %>% 
  mutate(mgr_tenure_group = cut(manager_tenure, c(0, 28, 45, 60, 90, 120, 180, 270, 365, 720, 99999))) %>%
  group_by(mgr_tenure_group, game_location, points) %>%
  tally() %>% 
  ungroup() %>%
  group_by(mgr_tenure_group, game_location) %>%
  mutate(n = n / sum(n)) %>%
  spread(points, n)

full_results %>% 
  mutate(mgr_tenure_group = cut(manager_tenure, c(-1, 28, 45, 60, 90, 120, 180, 270, 365, 720, 99999))) %>%
  group_by(mgr_tenure_group, game_location) %>%
  summarise(average_points = mean(points)) %>%
  ggplot(aes(x = mgr_tenure_group, y = average_points)) +
  geom_col() +
  facet_wrap(~ game_location)

goals <- full_result_data %>% 
  filter(valid_date) %>%
  add_count(game_id) %>% filter(n == 2) %>%
  pull(goals)

x <- MASS::fitdistr(goals, densfun = "Poisson") %>% broom::tidy()

exp_goals <- rpois(18000, lambda = x %>% pull(estimate))

curve(dpois(c(0:10), x %>% pull(estimate)))

plot(exp_goals)
```
