==> R/4-modelling.Rmd <==
---
title: "New Manager Bounce"
subtitle: "Modelling"
author: "Josh Muncke"
date: "21/03/2019"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(lme4)
```

## Modelling
```{r}
model_data <- game_and_manager_data %>%
  select(game_id, game_date, game_team, opponent, mgr_tenure_at_game, game_location, goals) %>%
  mutate(game_year = year(game_date),
         game_team = as.factor(game_team),
         opponent = as.factor(opponent),
         game_location = as.factor(game_location)) %>%
  select(-game_date)

model_data %<>% mutate_at(vars(game_year, mgr_tenure_at_game), scale)

model_train <- model_data %>% filter(game_year < 1.7)
model_test <- model_data %>% filter(game_year > 1.7)

goals_mod <- glmer(goals ~ game_location + game_year + (1 | game_team) + (1 | opponent) + mgr_tenure_at_game, 
                   data = model_train, 
                   family = poisson, 
                   verbose = 1)

model_test$preds <- round(exp(predict(goals_mod, newdata = model_test)),0)

model_test %>% 
  filter(game_location == 'Home') %>% 
  inner_join(model_test %>% filter(game_location == 'Away'), by = "game_id") %>% 
  mutate(real_result = case_when(goals.x > goals.y ~ "Home", goals.y > goals.x ~ "Away", TRUE ~ "Draw"),
         pred_result = case_when(preds.x > preds.y ~ "Home", preds.y > preds.x ~ "Away", TRUE ~ "Draw")) %>%
  count(real_result, pred_result) %>%
  mutate(n = n/sum(n)) %>%
  spread(pred_result, n)
```

```{r}
home_team <- "Manchester United"
away_team <- "Liverpool"

home_average <- game_and_manager_data %>%
  filter((game_team == home_team & game_location == "Home" & opponent == away_team)) %>%
  summarize(h_goals = mean(goals)) %>%
  pull(h_goals)


%>%
  group_by(goals) %>%
  summarise(n_games = n_distinct(game_id)) %>%
  mutate(pct_games = n_games / sum(n_games)) %>%
  ggplot() +
  geom_point(aes(x = goals, y = pct_games), colour = "red") +
  geom_point(aes(x = goals, y = rpois(goals, game_and_manager_data %>%
  filter((game_team == home_team & game_location == "Home" & opponent == away_team)) + summarise(mean_goals = mean(goals)) %>% pull(mean_goals) )), colour = "red") 


  ggplot(aes(x = goals)) +
  geom_histogram(binwidth = 1)
  
  
get_game_score <- function(home_team, away_team) {
  selected_subset <- game_and_manager_data %>%
    filter(game_date <= as.Date('2019-01-01')) %>%
    filter((game_team == home_team & game_location == "Home" & opponent == away_team)) 
  
  n_games <- selected_subset %>% summarise(n_games = n_distinct(game_id)) %>% pull(n_games)
  
  home_average <- selected_subset %>%
  summarize(avg_goals = mean(goals)) %>%
  pull(avg_goals)
  
  away_average <- selected_subset %>%
  summarize(avg_goals = mean(opponent_goals)) %>%
  pull(avg_goals)
  
  home_goals <- rpois(1, home_average)
  away_goals <- rpois(1, away_average) 
  
  tribble(~home, ~away, ~games,
          home_goals,away_goals,n_games)
}

get_game_score("Chelsea", "Crystal Palace")

crossing(trial = 1:1000,
         home_team = "Everton",
         away_team = "Chelsea") %>%
  mutate(result = map2(home_team, away_team, get_game_score)) %>%
  unnest() %>%
  mutate(result = if_else(home > away, home_team, if_else(home < away, away_team, "DRAW"))) %>%
  count(result) %>%
  mutate(n = n/sum(n)) %>%
  arrange(desc(n)) 
    
  

map(1:10, get_game_score("Chelsea", "Crystal Palace"))

rep(get_game_score("Bournemouth", "Newcastle"), 10)
x<-replicate(10, get_game_score("Chelsea", "Crystal Palace"))
```
